Created DB and Repos

+++ domain/step.go
@@ -1,9 +1,9 @@
 package domain
 
 type Step struct {
-	Id            int
+	Id            int64
 	PlanId        int
-	ReferenceId   int
+	ReferenceId   int64
 	ReferenceType ReferenceType
 	Position      int
 	Source        *Source
